refactor: split core traversal impls
